ON-5833: Move GHGI shared modules out of legacy inventory routes
