+++ app/src/app/[lng]/[inventory]/preferences/activities/page.tsx
@@ -2,10 +2,10 @@
 
 import React, { use } from "react";
 import { useTranslation } from "@/i18n/client";
-import PreferencesPageLayout from "../PreferencesPageLayout";
-import ActivitiesPage from "./ActivitiesPage";
+import PreferencesPageLayout from "@/components/GHGI/preferences/PreferencesPageLayout";
+import ActivitiesPage from "@/components/GHGI/preferences/ActivitiesPage";
 
-import { LINKS } from "@/app/[lng]/[inventory]/preferences/constants";
+import { LINKS } from "@/components/GHGI/preferences/constants";
 
 export default function ActivitiesLayout(props: {
   params: Promise<{ lng: string }>;
