+++ app/src/app/[lng]/[inventory]/preferences/transportation/page.tsx
@@ -2,10 +2,10 @@
 
 import React, { use } from "react";
 import { useTranslation } from "@/i18n/client";
-import PreferencesPageLayout from "../PreferencesPageLayout";
-import TransportationPage from "./TransportationPage";
+import PreferencesPageLayout from "@/components/GHGI/preferences/PreferencesPageLayout";
+import TransportationPage from "@/components/GHGI/preferences/TransportationPage";
 
-import { LINKS } from "@/app/[lng]/[inventory]/preferences/constants";
+import { LINKS } from "@/components/GHGI/preferences/constants";
 
 export default function TransportationLayout(props: {
   params: Promise<{ lng: string }>;
