+++ app/src/components/GHGI/inventory-result/ByActivityViewTable.tsx
