+++ app/src/components/GHGI/inventory-result/ByScopeView.tsx
