+++ app/src/components/GHGI/inventory-result/EmissionBySectorChart.tsx
