+++ app/src/components/GHGI/inventory-result/EmissionBySectorTable.tsx
