+++ app/src/components/GHGI/inventory-result/EmissionsForecast/CustomLegend.tsx
