+++ app/src/components/GHGI/inventory-result/EmissionsForecast/GrowthRatesExplanationModal.tsx
