+++ app/src/components/GHGI/inventory-result/EmissionsForecast/GrowthRatesExplanationModalTable.tsx
