+++ app/src/components/GHGI/inventory-result/EmissionsWidget.tsx
