+++ app/src/components/GHGI/inventory-result/EmptyStateCardContent.tsx
