+++ app/src/components/GHGI/inventory-result/SectorHeader.tsx
