+++ app/src/components/GHGI/inventory-result/index.tsx
