+++ app/src/components/GHGI/inventory-versions/VersionEntry.tsx
