+++ app/src/components/GHGI/inventory-versions/history-helpers.ts
