+++ app/src/components/GHGI/preferences/MinimalStepper.tsx
