+++ app/src/components/GHGI/preferences/PreferencesExplanation.tsx
