+++ app/src/components/Tabs/Activity/external-data-section.tsx
@@ -22,8 +22,8 @@ import { MdHomeWork } from "react-icons/md";
 import { getTranslationFromDict } from "@/i18n";
 import { DataCheckIcon } from "@/components/icons";
 import { FiCheckCircle, FiTarget } from "react-icons/fi";
-import { SourceDrawer } from "@/app/[lng]/[inventory]/data/[step]/SourceDrawer";
-import type { DataSourceWithRelations } from "@/app/[lng]/[inventory]/data/[step]/types";
+import { SourceDrawer } from "@/components/GHGI/data-step/SourceDrawer";
+import type { DataSourceWithRelations } from "@/components/GHGI/data-step/types";
 import { api } from "@/services/api";
 import { convertKgToTonnes } from "@/util/helpers";
 import { UseErrorToast } from "@/hooks/Toasts";
